更新 RatingStatisticsCard (評分 / 評分數 元件) (#1527)
