+++ src/components/CompanyAndJobTitle/StatisticsCard/StatisticsCard.module.css
@@ -1,0 +1,55 @@
+@value below-small, above-small from '../../common/variables.module.css';
+
+.statisticsCard {
+  display: flex;
+  width: fit-content;
+  margin-bottom: 12px;
+  border: 0.4px solid #cecece;
+  border-radius: 6px;
+  background-color: #fff;
+  box-shadow: 0 0 8px rgba(0, 0, 0, 0.13);
+}
+
+.statBlock {
+  position: relative;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  min-width: 55px;
+  padding: 8px 11px;
+  gap: 2px;
+
+  &::after {
+    content: '';
+    position: absolute;
+    top: 9px;
+    bottom: 9px;
+    right: 0;
+    width: 0.4px;
+    background-color: #cecece;
+  }
+
+  &:last-child::after {
+    display: none;
+  }
+
+  .label {
+    font-size: 11px;
+    font-weight: 400;
+  }
+
+  .value {
+    display: flex;
+    align-items: center;
+    gap: 2px;
+    font-size: 13px;
+    font-weight: 500;
+    min-height: 18px;
+
+    .blackThumb {
+      width: 6px;
+      height: 9.4px;
+      filter: brightness(0);
+    }
+  }
+}
