+++ src/components/common/OverallRating/Thumbs.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import ThumbImage from './thumb.svg';
+import ThumbImage from 'common/icons/thumb.svg';
 import cn from 'classnames';
 import styles from './Thumbs.module.css';
 import PropTypes from 'prop-types';
